refactor(sync): move pending sync to a separate module
